Add handleErrors option to avoid Gulp crashing on Stylus parse errors
This allows a user to avoid having to restart Gulp every time a syntax error is made and saved. Furthermore, it places the error cleanly into the CSS so that it appears in the browser. This ensures that when using LiveReload, users are not forced to constantly watch their gulp output while editing their styles.